+++ module-apps/application-antenna/ApplicationAntenna.hpp
@@ -50,7 +50,7 @@ namespace app
                            StartInBackground startInBackground = {false});
         virtual ~ApplicationAntenna();
 
-        sys::Message_t DataReceivedHandler(sys::DataMessage *msgl, sys::ResponseMessage *resp) override;
+        sys::MessagePointer DataReceivedHandler(sys::DataMessage *msgl, sys::ResponseMessage *resp) override;
         sys::ReturnCodes InitHandler() override;
         sys::ReturnCodes DeinitHandler() override;