+++ module-apps/application-calendar/ApplicationCalendar.cpp
@@ -52,7 +52,7 @@ namespace app
         busChannels.push_back(sys::BusChannels::ServiceDBNotifications);
     }
 
-    sys::Message_t ApplicationCalendar::DataReceivedHandler(sys::DataMessage *msgl, sys::ResponseMessage *resp)
+    sys::MessagePointer ApplicationCalendar::DataReceivedHandler(sys::DataMessage *msgl, sys::ResponseMessage *resp)
     {
         auto retMsg = Application::DataReceivedHandler(msgl);
         // if message was handled by application's template there is no need to process further.